requirements.txt
python-dotenv
tweepy
psycopg2
sqlalchemy
time
tweepy
neo4j
vaderSentiment
gensim
requests
alembic
translators
bertopic
```